refactor(playwright): co-locate playwright-ct with sanity package
