+++ .gitignore
@@ -71,9 +71,6 @@ etc
 /test/e2e/report
 /test/e2e/results
 /test/e2e/state
-/test/playwright-ct/report
-/test/playwright-ct/results
-/test/playwright-ct/template/.cache/
 
 # Temporary files made by yarn when running canary releases
 .yarn
